fix(integration): put the package in the quality gates, and fix what they found (#164)
`pyproject.toml` has listed `dotmac_integration.*` under mypy's strict settings
since the module landed. The Makefile never passed the package to `mypy` or
`bandit`. So the strictness was DECLARED and UNENFORCED — a config asserting a
standard nothing checked — and `0.1.0a1` published with 42 type errors and one
export that could never have run.

**`run_effect_once` raised `TypeError` on its first call.** It passed a
`payload=` keyword that `dotmac_kernel.idempotency.execute_once_platform` does
not accept, and an `operation` taking no arguments where the kernel calls
`operation(db)`. It is exported from `__init__`, and it has no caller and no
test — which is exactly why nothing noticed, and why the gate is the thing that
found it rather than a review. It is now a faithful adapter: every parameter is
the kernel's, the only addition is `mechanism` → `scope`, and the old `payload`
becomes `fingerprint`, the kernel's own column (ADR-0014).

The second real finding: `assert_connector_conforms` used `try/except/pass`
carrying `# noqa: S110` — ruff's code for that pattern, not bandit's `B110`. A
suppression aimed at the wrong tool, which is its own small lesson about
silencing by code rather than by reasoning. Rewritten as a positive assertion
with nothing swallowed.

The rest is annotation, in house style: 32 JSON-shaped payloads become
`dict[str, object]`. `secret_refs` is `dict[str, str]` downstream of validation
and `dict[str, object]` in the validator itself, which exists precisely to
reject non-strings — a distinction worth keeping rather than flattening.
`resolve_binding` types its rows once so three returns stop escaping as `Any`
from a declared return type. `_is_reference` narrows with `is not None` rather
than `bool(match)`: the old form was SAFE at runtime, since `and`
short-circuits, but a checker cannot prove it.

Version → `0.1.0a2`, since a published wheel exports the broken signature.

Verified by running the gates' exact commands: mypy over all 252 files clean,
bandit over the same set exits 0, ruff and import-linter clean.

+++ packages/dotmac-integration/src/dotmac_integration/retry.py
@@ -96,8 +96,10 @@ def retry_delay_seconds(
     exponent = max(attempt_count - 1, 0)
     # Guard the shift itself: 2 ** 10_000 is computed before min() sees it.
     if exponent > 32:
-        return policy.max_backoff_seconds
-    return min(policy.max_backoff_seconds, policy.base_delay_seconds * (2**exponent))
+        return int(policy.max_backoff_seconds)
+    return int(
+        min(policy.max_backoff_seconds, policy.base_delay_seconds * (2**exponent))
+    )
 
 
 def next_state(
